Tidy up update added file test
