+++ tests/Functional/Controller/SourcesControllerTest.php
@@ -14,7 +14,6 @@ use App\Enum\Source\Type;
 use App\Model\EntityId;
 use App\Repository\RunSourceRepository;
 use App\Repository\SourceRepository;
-use App\Request\AddFileRequest;
 use App\Request\FileSourceRequest;
 use App\Request\GitSourceRequest;
 use App\Request\InvalidSourceTypeRequest;
